Split stylesheet specific functions out of dom (#233873)
The dom file is getting pretty big. This splits out the parts that are specific to stylesheets as these are intendant of the other parts
