+++ src/vs/base/browser/domObservable.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { createStyleSheet2 } from './dom.js';
+import { createStyleSheet2 } from './domStylesheets.js';
 import { DisposableStore, IDisposable } from '../common/lifecycle.js';
 import { autorun, IObservable } from '../common/observable.js';
