+++ src/vs/platform/layout/browser/zIndexRegistry.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { clearNode, createCSSRule, createStyleSheet } from '../../../base/browser/dom.js';
+import { clearNode } from '../../../base/browser/dom.js';
+import { createCSSRule, createStyleSheet } from '../../../base/browser/domStylesheets.js';
 import { RunOnceScheduler } from '../../../base/common/async.js';
 
 export enum ZIndex {
