+++ src/vs/workbench/contrib/notebook/browser/diff/notebookDiffList.ts
@@ -6,6 +6,7 @@
 import './notebookDiff.css';
 import { IListMouseEvent, IListRenderer, IListVirtualDelegate } from '../../../../../base/browser/ui/list/list.js';
 import * as DOM from '../../../../../base/browser/dom.js';
+import * as domStylesheets from '../../../../../base/browser/domStylesheets.js';
 import { IListOptions, IListStyles, isMonacoEditor, IStyleController, MouseController } from '../../../../../base/browser/ui/list/listWidget.js';
 import { DisposableStore, IDisposable } from '../../../../../base/common/lifecycle.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
@@ -481,7 +482,7 @@ export class NotebookTextDiffList extends WorkbenchList<IDiffElementViewModelBas
 	override style(styles: IListStyles) {
 		const selectorSuffix = this.view.domId;
 		if (!this.styleElement) {
-			this.styleElement = DOM.createStyleSheet(this.view.domNode);
+			this.styleElement = domStylesheets.createStyleSheet(this.view.domNode);
 		}
 		const suffix = selectorSuffix && `.${selectorSuffix}`;
 		const content: string[] = [];
