+++ src/vs/workbench/contrib/webviewPanel/browser/webviewIconManager.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as dom from '../../../../base/browser/dom.js';
-import * as cssJs from '../../../../base/browser/cssValue.js';
+import * as cssValue from '../../../../base/browser/cssValue.js';
+import * as domStylesheets from '../../../../base/browser/domStylesheets.js';
 import { Disposable } from '../../../../base/common/lifecycle.js';
 import { URI } from '../../../../base/common/uri.js';
 import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
@@ -39,7 +39,7 @@ export class WebviewIconManager extends Disposable {
 
 	private get styleElement(): HTMLStyleElement {
 		if (!this._styleElement) {
-			this._styleElement = dom.createStyleSheet(undefined, undefined, this._store);
+			this._styleElement = domStylesheets.createStyleSheet(undefined, undefined, this._store);
 			this._styleElement.className = 'webview-icons';
 		}
 		return this._styleElement;
@@ -67,8 +67,8 @@ export class WebviewIconManager extends Disposable {
 				const webviewSelector = `.show-file-icons .webview-${key}-name-file-icon::before`;
 				try {
 					cssRules.push(
-						`.monaco-workbench.vs ${webviewSelector}, .monaco-workbench.hc-light ${webviewSelector} { content: ""; background-image: ${cssJs.asCSSUrl(value.light)}; }`,
-						`.monaco-workbench.vs-dark ${webviewSelector}, .monaco-workbench.hc-black ${webviewSelector} { content: ""; background-image: ${cssJs.asCSSUrl(value.dark)}; }`
+						`.monaco-workbench.vs ${webviewSelector}, .monaco-workbench.hc-light ${webviewSelector} { content: ""; background-image: ${cssValue.asCSSUrl(value.light)}; }`,
+						`.monaco-workbench.vs-dark ${webviewSelector}, .monaco-workbench.hc-black ${webviewSelector} { content: ""; background-image: ${cssValue.asCSSUrl(value.dark)}; }`
 					);
 				} catch {
 					// noop
